color improvements

+++ src/lib/ecs/traits.ts
@@ -27,7 +27,7 @@ export const Color = trait({ r: 0, g: 0, b: 0 })
 export const Arrow = trait()
 
 export const Positions = trait(() => new Float32Array())
-export const Colors = trait(() => new Float32Array())
+export const Colors = trait(() => new Uint8Array())
 export const Instances = trait({
 	ids: () => new Uint32Array(),
 	meshId: -1,
